Fixed OnStart PC Case Tilt
Fixed the issue with the PC case Tilting over due to the physics effects of the cover attaching to the Socket. There is a Script called StopStartUpCasingTiltOver.cs that does the job. It does a rigidbody  constraint at statup and release it after 5 seconds. The also used On Hover to turn it ensure case follows phsics